+++ src/main/java/com/blocklegend001/immersiveores/util/tools/hammer/HammerOverlayRenderer.java
@@ -4,15 +4,19 @@ import com.blocklegend001.immersiveores.item.ModItems;
 import com.blocklegend001.immersiveores.item.custom.base.Hammer;
 import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderEvents;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.render.*;
+import net.minecraft.client.render.RenderLayer;
+import net.minecraft.client.render.VertexConsumer;
+import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import net.minecraft.registry.tag.BlockTags;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.hit.HitResult;
-import net.minecraft.util.math.*;
-import net.minecraft.item.ItemStack;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Box;
 import net.minecraft.util.math.Direction;
+import net.minecraft.util.math.Vec3d;
 import org.joml.Matrix4f;
 
 import java.util.Map;
@@ -37,7 +41,12 @@ public class HammerOverlayRenderer {
 
             Direction side = blockHit.getSide();
             BlockPos origin = blockHit.getBlockPos();
-            int range = HAMMER_RANGERS.getOrDefault(heldItem.getItem(), 1);
+            int range;
+            if (client.player.isSneaking()) {
+                range = 0;
+            } else {
+                range = HAMMER_RANGERS.get(heldItem.getItem());
+            }
 
             if (!client.world.getBlockState(origin).isIn(BlockTags.PICKAXE_MINEABLE)) return;